Rename ld-comb-pal to ld-chroma-decoder.
This is in preparation for merging the NTSC and PAL decoders.
