+++ tools/.gitignore
@@ -43,7 +43,7 @@ target_wrapper.*
 # QtCreator CMake
 CMakeLists.txt.user*
 /ld-analyse/ld-analyse
-/ld-comb-pal/ld-comb-pal
+/ld-chroma-decoder/ld-chroma-decoder
 /ld-dropout-correct/ld-dropout-correct
 /ld-process-vbi/ld-process-vbi
 /ld-process-ntsc/ld-process-ntsc
