+++ tools/ld-decode-tools.pro
@@ -1,9 +1,9 @@
 TEMPLATE = subdirs
 SUBDIRS = \
     ld-analyse \
+    ld-chroma-decoder \
     ld-comb-ntsc \
     ld-comb-ntsc/testfilter \
-    ld-comb-pal \
     ld-combine \
     ld-decode-shared \
     ld-dropout-correct \
@@ -14,8 +14,8 @@ SUBDIRS = \
     ld-process-vbi
 
 ld-analyse.depends = ld-decode-shared
+ld-chroma-decoder.depends = ld-decode-shared
 ld-comb-ntsc.depends = ld-decode-shared
-ld-comb-pal.depends = ld-decode-shared
 ld-combine.depends = ld-decode-shared
 ld-dropout-correct.depends = ld-decode-shared
 ld-process-vbi.depends = ld-decode-shared
